requirements.txt
pandas==2.2.2
Pillow==10.3.0
plotly==5.20.0
streamlit==1.35.0
